requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
tensorflow==2.16.1
numpy==1.26.4
pillow==10.3.0
pandas==2.1.4
joblib==1.3.2
xgboost==2.0.3